backend/generate-questions/requirements.txt
openai==1.54.3
psycopg2-binary==2.9.9